feature: add price tables controllers

+++ colanta-backend/App/PriceTables/FixedPrice.cs
@@ -5,7 +5,7 @@ namespace colanta_backend.App.PriceTables{
         public int Id { get; set; }
         public decimal Value { get; set; }
         public decimal ListPrice { get; set; }
-        public decimal VtexSkuId { get; set; }
+        public int VtexSkuId { get; set; }
         public string PriceTableName { get; set; }
         public PriceTable PriceTable { get; set; } 
         public int MinQuantity { get; set; }
